+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.23.1] - 2024-05-17
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Option to `TopBar` dropdown.
+- [Marek Kurańda](https://github.com/mjkuranda): View for `/transactions` page.
+
 ## [0.23.0] - 2024-05-16
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): `/transactions` page to render all user's transactions.